refactor: split Typecheck to reduce dependencies on TypeError
