+++ primer/src/Primer/Questions.hs
@@ -32,7 +32,8 @@ import Primer.Def (
 import Primer.Name (Name, unName, unsafeMkName)
 import Primer.Name.Fresh (mkAvoidForFreshName, mkAvoidForFreshNameTy, mkAvoidForFreshNameTypeZ)
 import Primer.TypeDef (typeDefNameHints)
-import Primer.Typecheck (Cxt, decomposeTAppCon, typeDefs)
+import Primer.Typecheck.Cxt (Cxt, typeDefs)
+import Primer.Typecheck.Utils (decomposeTAppCon)
 import Primer.Zipper (
   ExprZ,
   TypeZ,
