+++ primer/src/Primer/Typecheck.hs
@@ -61,21 +61,17 @@ module Primer.Typecheck (
 
 import Foreword
 
-import Control.Arrow ((&&&))
 import Control.Monad.Fresh (MonadFresh (..))
 import Control.Monad.NestedError (MonadNestedError (..))
-import Data.Functor.Compose (Compose (Compose), getCompose)
 import Data.Map qualified as M
 import Data.Map.Strict qualified as Map
 import Data.Set qualified as S
-import Data.Tuple.Extra (fst3)
 import Optics (
   A_Traversal,
   AppendIndices,
   IxFold,
   IxTraversal',
   JoinKinds,
-  Lens',
   Optic',
   WithIx,
   equality,
@@ -88,7 +84,6 @@ import Optics (
   set,
   to,
   traverseOf,
-  view,
   (%),
  )
 import Optics.Traversal (traversed)
@@ -105,7 +100,6 @@ import Primer.Core (
   LVarName,
   LocalName (LocalName),
   Meta (..),
-  ModuleName,
   PrimCon,
   TmVarRef (..),
   TyConName,
@@ -120,9 +114,7 @@ import Primer.Core (
   unLocalName,
   _bindMeta,
   _exprMeta,
-  _exprMetaLens,
   _exprTypeMeta,
-  _type,
   _typeMeta,
  )
 import Primer.Core.DSL (branch, emptyHole, meta, meta')
@@ -157,11 +149,10 @@ import Primer.TypeDef (
   TypeDefMap,
   ValCon (valConArgs, valConName),
   typeDefAST,
-  typeDefParameters,
   valConType,
  )
+import Primer.Typecheck.Cxt (Cxt (Cxt, globalCxt, localCxt, smartHoles, typeDefs))
 import Primer.Typecheck.Kindcheck (
-  Cxt (Cxt, globalCxt, localCxt, smartHoles, typeDefs),
   KindError (..),
   KindOrType (K, T),
   Type,
@@ -176,6 +167,23 @@ import Primer.Typecheck.Kindcheck (
  )
 import Primer.Typecheck.SmartHoles (SmartHoles (..))
 import Primer.Typecheck.TypeError (TypeError (..))
+import Primer.Typecheck.Utils (
+  TypeDefError (TDIHoleType, TDINotADT, TDINotSaturated, TDIUnknownADT),
+  TypeDefInfo (TypeDefInfo),
+  decomposeTAppCon,
+  getGlobalBaseNames,
+  getGlobalNames,
+  getTypeDefInfo,
+  getTypeDefInfo',
+  instantiateValCons,
+  instantiateValCons',
+  lookupConstructor,
+  maybeTypeOf,
+  mkTAppCon,
+  substituteTypeVars,
+  typeOf,
+  _typecache,
+ )
 
 -- | Typechecking takes as input an Expr with 'Maybe Type' annotations and
 -- produces an Expr with 'Type' annotations - i.e. every node in the output is
@@ -260,18 +268,6 @@ type TypeM e m =
   , MonadFresh NameCounter m
   , MonadNestedError TypeError e m -- can throw type errors
   )
-
--- | A lens for the type annotation of an 'Expr' or 'ExprT'
-_typecache :: Lens' (Expr' (Meta a) b) a
-_typecache = _exprMetaLens % _type
-
--- | Get the (potentially absent) type of an 'Expr'
-maybeTypeOf :: Expr -> Maybe TypeCache
-maybeTypeOf = view _typecache
-
--- | Get the type of an 'ExprT'
-typeOf :: ExprT -> TypeCache
-typeOf = view _typecache
 
 -- | Check a context is valid
 checkValidContext ::
@@ -432,15 +428,6 @@ checkEverything sh CheckEverything{trusted, toCheck} =
         traverseDefs' (reindexed moduleName selfIndex)
           % to defType
           % to forgetTypeMetadata
-
--- We assume that constructor names are unique, returning the first one we find
-lookupConstructor :: TypeDefMap -> ValConName -> Maybe (ValCon, TyConName, ASTTypeDef)
-lookupConstructor tyDefs c =
-  let allCons = do
-        (tc, TypeDefAST td) <- M.assocs tyDefs
-        vc <- astTypeDefConstructors td
-        pure (vc, tc, td)
-   in find ((== c) . valConName . fst3) allCons
 
 {- HLINT ignore synth "Avoid lambda using `infix`" -}
 -- Note [Let expressions]
@@ -718,72 +705,6 @@ check t = \case
             e'' -> pure e''
       _ -> default_
 
-data TypeDefError
-  = TDIHoleType -- a type hole
-  | TDINotADT -- e.g. a function type etc
-  | TDIUnknownADT TyConName -- not in scope
-  | TDINotSaturated -- e.g. @List@ or @List a b@ rather than @List a@
-
-data TypeDefInfo a = TypeDefInfo [Type' a] TyConName TypeDef -- instantiated parameters, and the typedef (with its name), i.e. [Int] are the parameters for @List Int@
-
-getTypeDefInfo :: MonadReader Cxt m => Type' a -> m (Either TypeDefError (TypeDefInfo a))
-getTypeDefInfo t = reader $ flip getTypeDefInfo' t . typeDefs
-
-getTypeDefInfo' :: TypeDefMap -> Type' a -> Either TypeDefError (TypeDefInfo a)
-getTypeDefInfo' _ (TEmptyHole _) = Left TDIHoleType
-getTypeDefInfo' _ (THole _ _) = Left TDIHoleType
-getTypeDefInfo' tydefs ty =
-  case decomposeTAppCon ty of
-    Nothing -> Left TDINotADT
-    Just (tycon, params) -> do
-      case M.lookup tycon tydefs of
-        Nothing -> Left $ TDIUnknownADT tycon
-        Just tydef
-          -- this check would be redundant if we were sure that the input type
-          -- were of kind KType, alternatively we should do kind checking here
-          | length (typeDefParameters tydef) /= length params -> Left TDINotSaturated
-          | otherwise -> Right $ TypeDefInfo params tycon tydef
-
--- | Takes a particular instance of a parameterised type (e.g. @List Nat@), and
--- extracts both both the raw typedef (e.g. @List a = Nil | Cons a (List a)@)
--- and the constructors with instantiated argument types
--- (e.g. @Nil : List Nat ; Cons : Nat -> List Nat -> List Nat@)
-instantiateValCons ::
-  (MonadFresh NameCounter m, MonadReader Cxt m) =>
-  Type' () ->
-  m (Either TypeDefError (TyConName, ASTTypeDef, [(ValConName, [Type' ()])]))
-instantiateValCons t = do
-  tds <- asks typeDefs
-  let instCons = instantiateValCons' tds t
-      -- Because @(,,) a b@ does not have a Traversable instance
-      -- we reassociate so we use the one of @(,) a@
-      reassoc (a, b, c) = ((a, b), c)
-      reassoc' ((a, b), c) = (a, b, c)
-      sequence4 =
-        fmap (getCompose . getCompose . getCompose . getCompose)
-          . sequence
-          . Compose
-          . Compose
-          . Compose
-          . Compose
-  fmap (fmap reassoc') $ sequence4 $ fmap reassoc instCons
-
--- | As 'instantiateValCons', but pulls out the relevant bits of the monadic
--- context into an argument
-instantiateValCons' ::
-  MonadFresh NameCounter m =>
-  TypeDefMap ->
-  Type' () ->
-  Either TypeDefError (TyConName, ASTTypeDef, [(ValConName, [m (Type' ())])])
-instantiateValCons' tyDefs t = do
-  TypeDefInfo params tc def <- getTypeDefInfo' tyDefs t
-  case def of
-    TypeDefPrim _ -> Left TDINotADT
-    TypeDefAST tda -> do
-      let defparams = map fst $ astTypeDefParameters tda
-          f c = (valConName c, map (substituteTypeVars $ zip defparams params) $ valConArgs c)
-      pure (tc, tda, map f $ astTypeDefConstructors tda)
-
 -- | Similar to check, but for the RHS of case branches
 -- We assume that the branch is for this constructor
 -- The passed in 'ValCon' is assumed to be a constructor of the passed in 'TypeDef'
@@ -827,25 +748,6 @@ checkBranch t (vc, args) (CaseBranch nb patterns rhs) =
           <> show vc
           <> " but found branch on "
           <> show nb
-
-substituteTypeVars :: MonadFresh NameCounter m => [(TyVarName, Type' ())] -> Type' () -> m (Type' ())
-substituteTypeVars = flip $ foldrM (uncurry substTy)
-
--- | Decompose @C X Y Z@ to @(C,[X,Y,Z])@
-decomposeTAppCon :: Type' a -> Maybe (TyConName, [Type' a])
-decomposeTAppCon ty = do
-  (con, args) <- go ty
-  pure (con, reverse args)
-  where
-    go (TCon _ con) = Just (con, [])
-    go (TApp _ t s) = do
-      (con, args) <- go t
-      pure (con, s : args)
-    go _ = Nothing
-
--- | @mkTAppCon C [X,Y,Z] = C X Y Z@
-mkTAppCon :: TyConName -> [Type' ()] -> Type' ()
-mkTAppCon c = foldl' (TApp ()) (TCon () c)
 
 -- | Checks if a type can be unified with a function (arrow) type. Returns the
 -- arrowised version - i.e. if it's a hole then it returns an arrow type with
@@ -904,28 +806,6 @@ exprTtoExpr = over _exprTypeMeta (fmap Just) . over _exprMeta (fmap Just)
 -- | Convert @Type (Meta Kind)@ to @Type (Meta (Maybe Kind))@
 typeTtoType :: TypeT -> Type' TypeMeta
 typeTtoType = over _typeMeta (fmap Just)
-
--- Helper to create fresh names
-getGlobalNames :: MonadReader Cxt m => m (S.Set (ModuleName, Name))
-getGlobalNames = do
-  tyDefs <- asks typeDefs
-  topLevel <- asks $ S.fromList . map f . M.keys . globalCxt
-  let ctors =
-        Map.foldMapWithKey
-          ( \t def ->
-              S.fromList $
-                (f t :) $
-                  map (f . valConName) $
-                    maybe [] astTypeDefConstructors $
-                      typeDefAST def
-          )
-          tyDefs
-  pure $ S.union topLevel ctors
-  where
-    f = qualifiedModule &&& baseName
-
-getGlobalBaseNames :: MonadReader Cxt m => m (S.Set Name)
-getGlobalBaseNames = S.map snd <$> getGlobalNames
 
 checkKind' :: TypeM e m => Kind -> Type' (Meta a) -> m TypeT
 checkKind' k t =
